requirements.txt
tensorflow
tensorflow-gpu==1.9
numpy
scipy
Pillow
matplotlib
